Refactor `AwsSecretsManagerAtlassianTokenStore` into `AtlassianTokenStore` for simplifying architecture and removing unnecessary interface abstraction in `token-refactor`.
